% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_variable_name}
\alias{create_variable_name}
\title{Create variable name from string.}
\usage{
create_variable_name(s)
}
\arguments{
\item{s}{The string to be used as a variable name}
}
\value{
The string with non-alphanumeric characters replaced by underscores
}
\description{
This function creates a sensible variable name by replacing
all non-alphanumeric characters with underscores. This is useful
when the string is to be used as a the name of a list element.
}
\keyword{internal}

