% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dec2hex}
\alias{dec2hex}
\title{Convert a decimal number to a hexadecimal string.}
\usage{
dec2hex(x)
}
\arguments{
\item{x}{A decimal number.}
}
\value{
A hexadecimal string.
}
\description{
Convert a decimal number to a hexadecimal string.
}
\keyword{internal}

