% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ebm.R
\docType{package}
\name{edf}
\alias{edf}
\alias{edf-package}
\title{Read data stored in the Embla Data Format (EBM).}
\description{
The ebm package provides methods for reading files containing
physiologic data stored in the Embla Data Format (EBM).
}

