% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_data_files}
\alias{get_data_files}
\title{Return a list of EBM signal files.}
\usage{
get_data_files(datadir, pattern = NULL)
}
\arguments{
\item{datadir}{The directory containing the EBM files.}

\item{pattern}{A pattern used to find a signal file.}
}
\value{
The data files as a list.
}
\description{
Return a list of EBM signal files.
}
\keyword{internal}

