% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ebm_record_type}
\alias{get_ebm_record_type}
\title{A list of known record fields in the EBM file.}
\usage{
get_ebm_record_type(x)
}
\arguments{
\item{x}{A hexadecimal string (without "0x").x}
}
\value{
A string describing the header record.
}
\description{
A list of known record fields in the EBM file.
}
\keyword{internal}

