% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{hex2dec}
\alias{hex2dec}
\title{Convert a hexadecimal string to decimal.}
\usage{
hex2dec(x)
}
\arguments{
\item{x}{A hexadecimal string.}
}
\value{
A decimal number.
}
\description{
Convert a hexadecimal string to decimal.
}
\keyword{internal}

