% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{new_recording}
\alias{new_recording}
\title{#' Return and initialise an empty recording structure.
The recording structure is a list.}
\usage{
new_recording()
}
\value{
An empty recording structure.
}
\description{
#' Return and initialise an empty recording structure.
The recording structure is a list.
}

