% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_ebm_record}
\alias{parse_ebm_record}
\title{Parse an EBM record, given the record id as a string and the raw record data.}
\usage{
parse_ebm_record(record, data, endian = "big")
}
\arguments{
\item{record}{A string describing the record.}

\item{data}{The raw bytes contained in the record.}

\item{endian}{The endianness. Defaults to big.}
}
\value{
The parsed record data (number of string).
}
\description{
Parse an EBM record, given the record id as a string and the raw record data.
}
\keyword{internal}

