% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ebm.R
\name{read.ebm}
\alias{read.ebm}
\title{Read data in the Embla Data Format.}
\usage{
read.ebm(datapath, channels = NULL, start = 0, data.length = NULL,
  header.only = FALSE)
}
\arguments{
\item{datapath}{Either the full path to a directory containing
EBM-files or the full path to a particular EBM file. If a directory
is given, it is assumed that all these files are from the same
recording session}

\item{channels}{A string or a list of strings with channel names
(typically the same name as the EBM file names). If the
\code{datapath} argument is a directory, these strings are used to
filter files from the directory.}

\item{start}{The offset in seconds from the beginning of the file to start reading data.}

\item{data.length}{The amount of data in seconds to be read.}

\item{header.only}{Boolean denoting whether to only read the headers in the EBM file. Default is FALSE.}
}
\value{
The data in the EBM file(s) as a list.
}
\description{
This function reads the data stored in an EBM file. This data
consists of, e.g., physiologic signals.
}
\examples{
\dontrun{
# Read all EBM files in a directory
ebm <- read.ebm("/tmp/datapath")

# Read one particular EBM file
ebm <- read.ebm("/tmp/datapath/ECG.ebm")

# Read some channels
ebm <- read.ebm("/tmp/datapath", channels = c("ECG", "Fz", "Cz"))

# Read 20 seconds of data, starting 10 seconds after the beginning of the file:
ebm <- read.ebm("/tmp/datapath/ECG.ebm", start = 10, data.length = 20)

# Only read the header
ebm <- read.ebm("/tmp/datapath/ECG.ebm", header.only = TRUE)
}

}

