% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ebm.R
\name{read.ebm.single}
\alias{read.ebm.single}
\title{Read data in Embla Data Format from a single file.}
\usage{
read.ebm.single(datafile, start = 0, data.length = NULL,
  header.only = FALSE)
}
\arguments{
\item{datafile}{The full path to a particular EBM file.}

\item{start}{The offset in seconds from the beginning of the file to start reading data.}

\item{data.length}{The amount of data in seconds to be read.}

\item{header.only}{Boolean denoting whether to only read the headers in the EBM file. Default is FALSE.}
}
\value{
The data in the EBM file as a list.
}
\description{
This function reads the data stored in an EBM file. This data
consists of, e.g., physiologic signals.
}
\keyword{internal}

